+++ lib/features/home/screens/home_screen.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../../core/theme/app_theme.dart';
 import '../../../core/constants/colors.dart';
-import '../providers/home_providers.dart';
 import '../widgets/next_prayer_card.dart';
 import '../widgets/today_prayers_list.dart';
 import '../widgets/monthly_stats_card.dart';